+++ R/load.R
@@ -40,6 +40,9 @@ load <- function(project = NULL, quiet = FALSE) {
   project <- project %||% renv_project_find(project)
   renv_scope_lock(project = project)
 
+  # avoid suppressing the next auto snapshot
+  renv_scope_var("running", TRUE, envir = `_renv_snapshot_auto`)
+
   # if load is being called via the autoloader,
   # then ensure RENV_PROJECT is unset
   # https://github.com/rstudio/renv/issues/887