+++ R/project.R
@@ -262,6 +262,9 @@ renv_project_synchronized_check <- function(project = NULL, lockfile = NULL) {
   # to read and use it as-is
   renv_scope_options(renv.config.snapshot.validate = FALSE)
 
+  # indicate that we're doing an automatic snapshot
+  renv_scope_var("running", TRUE, envir = `_renv_auto_snapshot`)
+
   # read current state of project library
   libstate <- snapshot(
     project  = project,